Simulations

The Simulation object is at the heart of WAVI.jl. (An instance of) a Simulation is constructed by passing an instance of a Model, an instance of a TimesteppingParams object, and (optionally) an instance of an OutputParams object, e.g.

simulation = Simulation(model = model, output_params = output_params, timestepping_params = timestepping_params)

WAVI.Simulation — Method

Simulation(;
        model = nothing,
        timestepping_params = nothing,
        output_params = OutputParams(),
        pickup_output_update_flag = false)

Construct a WAVI.jl Simulation object.

Keyword arguments

- `model`: (required) an instance of a `Model`` object
- `timestepping_params`: (required) an instance of a `TimesteppingParams` object, which stores information relating to timestepping
- `output_params`: an instance of an `OutputParams` object, which stores information relating to outputting of solutions
- `pickup_output_update_flag`: a flag which specifies whether to update the output_params upon picking up.